requirements.txt
celery==5.1.2
coverage==6.1.1
cryptography==35.0.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask==2.0.2
gunicorn==20.1.0
PyMySQL==1.0.2
pytest==6.2.5
requests==2.26.0
Werkzeug==2.0.2
